// management/frontend/src/api/schemas/menuTreeNode.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface MenuTreeNode {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  parent_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  type?: number;
  path?: string | null;
  icon?: string | null;
  permission_code?: string | null;
  sort?: number;
  is_active?: boolean;
  is_visible?: boolean;
  is_cache?: boolean;
  creator_name?: string | null;
  updater_name?: string | null;
  children?: MenuTreeNode[];
}